content-ai/requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt==4.0.1 
passlib==1.7.4
python-multipart
python-dotenv
alembic

langchain
langchain-openai
langchain-community

